+++ project/OsEngine/Market/Servers/Tester/TesterServer.cs
@@ -4280,7 +4280,7 @@ namespace OsEngine.Market.Servers.Tester
 
             if (NewBidAskIncomeEvent != null)
             {
-                NewBidAskIncomeEvent((double)candle.Close, (double)candle.Close, GetSecurityForName(nameSecurity, ""));
+                NewBidAskIncomeEvent((decimal)candle.Close, (decimal)candle.Close, GetSecurityForName(nameSecurity, ""));
             }
 
             _candleManager.SetNewCandleInSeries(candle, nameSecurity, timeFrame);
@@ -4313,7 +4313,7 @@ namespace OsEngine.Market.Servers.Tester
 
         public event Action<MarketDepth> NewMarketDepthEvent;
 
-        public event Action<double, double, Security> NewBidAskIncomeEvent;
+        public event Action<decimal, decimal, Security> NewBidAskIncomeEvent;
 
         #endregion
 
@@ -4409,7 +4409,7 @@ namespace OsEngine.Market.Servers.Tester
             }
             if (NewBidAskIncomeEvent != null)
             {
-                NewBidAskIncomeEvent((double)tradesNew[tradesNew.Count - 1].Price, (double)tradesNew[tradesNew.Count - 1].Price, GetSecurityForName(tradesNew[tradesNew.Count - 1].SecurityNameCode, ""));
+                NewBidAskIncomeEvent((decimal)tradesNew[tradesNew.Count - 1].Price, (decimal)tradesNew[tradesNew.Count - 1].Price, GetSecurityForName(tradesNew[tradesNew.Count - 1].SecurityNameCode, ""));
             }
         }
